lambdas added

+++ src/main/java/offer/Bogof.java
@@ -5,9 +5,9 @@ package offer;
  */
 public class Bogof implements SpecialOffer {
 
-    public double applyDiscount(int numberOfItems, double price) {
+    public double applyDiscount(long numberOfItems, double price) {
         if(numberOfItems > 0) {
-            int rest = numberOfItems % 2;
+            long rest = numberOfItems % 2;
             return ((numberOfItems - rest) / 2 + rest) * price;
         } else {
             return 0;
