+++ src/main/java/offer/ThreeForTwo.java
@@ -5,9 +5,9 @@ package offer;
  */
 public class ThreeForTwo implements SpecialOffer {
 
-    public double applyDiscount(int numberOfItems, double price) {
+    public double applyDiscount(long numberOfItems, double price) {
         if(numberOfItems > 0) {
-            int rest = numberOfItems % 3;
+            long rest = numberOfItems % 3;
             return ((numberOfItems - rest) / 3) * price * 2 + (rest * price);
         } else {
             return 0;
